See desc.
* Moved the pointers to various features of aRAMa that are allocated into dynamic into a struct
* More work done on the CodeHandler
* Beginning work on a notification/logging manager class, that will do both logging and notifications simultaneously
* Moved the NodeList class into a dedicated file
* Created a base ThreadServer class for running, well, threaded servers
* A lot of work towards making the CommandHandler work again

+++ src/gecko/comm_hand_backend/comm_write_int.cpp
@@ -33,12 +33,12 @@ void CommandHandler::command_write_16()
 
 void CommandHandler::command_write_32()
 {
-	int destinationAddress, value;
+	uint8_t destinationAddress, value;
 	ret = recvwait(sizeof(int) * 2);
 	CHECK_ERROR(ret < 0)
 
-	destinationAddress = ((int *)buffer)[0];
-	value = ((int *)buffer)[1];
+	destinationAddress = ((uint8_t *) buffer)[0];
+	value = ((uint8_t *)buffer)[1];
 
 	GeckoKernelCopyData((uint8_t *) destinationAddress, (uint8_t *) &value, sizeof(int));
